working example of best fit objective; added compound image reporter and image grid reporter; improved Monitor/Feed construction; added version to __init__; added TailingDataFeed; separated updates into model updates and tap updates

+++ src/objective.py
@@ -60,7 +60,7 @@ class BestFit(Objective):
         Returns an initial value suitable for the shared variable which will be used by taps
         to capture the value of the expected value.
         '''
-        return np.zeros((self.layer.batch_size, self.layer.size), dtype=np.int64)
+        return np.zeros((self.layer.batch_size, self.layer.size), dtype=np.single)
 
 
     def output(self):
